A whole bunch of stuff was done here, I'll name what I can. Fixed monster_chick (was monster_bitch) Fixed a delay problem with func_timers Fixed objects disappearing into certain objects and railgun going through certain objects Moved a whole bunch of things out of edict_s and into their proper classes Fixed intermission origins Tons of vec3_t->vec3f changes Added a frame-based offset to the parasite cable (it pretty much always comes from above his head now in the right spot instead of always in the same spot) A few memory issues were resolved. Release version can compile now. Found a problem in Mutamt's touch function that might of rendered them useless after one jump; fixed that.
Notes:
Firing a railgun that starts inside Makron's body goes into an infinite loop and never returns. This seems to be because when you fire a railgun that starts inside a solid body, instead of hitting once and ignoring it, it keeps hitting the body; on any old body, the body will just gib and you won't notice it, but Makron's body never gibs, so this might be a problem.
Changing levels, saving and loading will require a specially made system that I don't feel like doing right now. I'll think of something.